refactored routes to fail louldy and to include user id

+++ server/routes/tasks.js
@@ -14,7 +14,7 @@ router.get("/", async (req, res) => {
 });
 
 //get a single task 
-const get_single_task = router.get("/:id", async (req, res) => {
+router.get("/:id", async (req, res) => {
   try {
     const { id } = req.params;
     const task = await pool.query(
@@ -29,67 +29,80 @@ const get_single_task = router.get("/:id", async (req, res) => {
   }
 });
 
-//create task
+//create task as user
 router.post("/create", async (req, res) => {
   try {
     //description: varchar(255), complete: boolean
-    const { task_description, complete } = req.body;
+    const { app_user_id, task_description, complete } = req.body;
     const new_task = await pool.query(
-      "INSERT INTO tasks (task_description, complete) VALUES ($1, $2) returning *",
-      [task_description, complete]
+      "INSERT INTO tasks (app_user_id, task_description, complete) VALUES ($1, $2, $3) RETURNING *",
+      [app_user_id, task_description, complete]
       );
 
     //respond with newly created task
-    res.json(new_task.row[0]);
+    res.json(new_task.rows[0]);
 
   } catch (error) {
     console.log(error.message);
   }
 });
 
-//update task
-router.put("/:id", async (req, res) => {
+//update task as user that owns task
+router.put("/:app_user_id/:task_id", async (req, res) => {
   try {
     //description: varchar(255), complete: boolean
     const { task_description, complete } = req.body;
-    const { id } = req.params;
-    const update_task = pool.query(
-      "UPDATE tasks SET task_description = $1, complete = $2 WHERE task_id = $3",
-      [task_description, complete, id]
+    const { app_user_id, task_id } = req.params;
+    const update_task = await pool.query(
+      "UPDATE tasks SET task_description = $1, complete = $2 WHERE task_id = $3 AND app_user_id = $4 RETURNING *",
+      [task_description, complete, task_id, app_user_id]
     );
 
-    res.json("Task successfully updated");
+    if (!!update_task.rows[0] === true) {
+      res.json("Task successfully updated")
+    } else {
+      res.json("Failed to update task")
+    }
 
   } catch (error) {
     console.log(error.message);
   }
 });
 
-//delete task
-router.delete("/:id", async (req, res) => {
+//delete task as user that owns task
+router.delete("/:app_user_id/:task_id", async (req, res) => {
   try {
-    const { id } = req.params;
-    const delete_task = pool.query(
-      "DELETE FROM tasks where task_id = ($1)",
-      [id]
+    const { app_user_id, task_id } = req.params;
+    const delete_task = await pool.query(
+      "DELETE FROM tasks where task_id = $1 AND app_user_id = $2 RETURNING *",
+      [task_id, app_user_id]
     );
 
-    res.json("Task successfully deleted");
+    if (!!delete_task.rows[0] === true) {
+      res.json("Task successfully deleted")
+    } else {
+      res.json("Failed to delete task")
+    }
+
   } catch (error) {
     console.log(error.message);
   }
 });
 
-//toggle task completion
-router.put("/complete/:id", async (req, res) => {
+//toggle task completion as owner of task 
+router.put("/complete/:app_user_id/:task_id", async (req, res) => {
   try {
-    const { id } = req.params;
-    const toggle_copmlete = pool.query(
-      "UPDATE tasks SET complete = NOT complete WHERE task_id = $1",
-      [id]
+    const { app_user_id, task_id } = req.params;
+    const toggle_complete = await pool.query(
+      "UPDATE tasks SET complete = NOT complete WHERE task_id = $1 AND app_user_id = $2 RETURNING *",
+      [task_id, app_user_id]
     );
-
-    res.json("Task completion successfully toggled");
+    
+    if (!!toggle_complete.rows[0] === true) {
+      res.json("Task completion successfully toggled")
+    } else {
+      res.json("Task failed to toggle")
+    }
 
   } catch (error) {
     console.log(error.message);
